Convolutional Neural Network for the CIFAR10 datset
This is my first CNN, and it is trained and tested in the code. I still have to build functions to test and train models, however, this was a first step. I can also make a more developed structure for the neural network, as this one is pretty standard. The bulk of the operations are in the main.py function, and the code for the structure of the Neural Network is in the NNClass.py file.